Add middleware auth

+++ internal/domain/user.go
@@ -3,8 +3,8 @@ package domain
 import "time"
 
 var (
-	ITRole    = 615
-	NurseRole = 303
+	ITRole    = int64(615)
+	NurseRole = int64(303)
 )
 
 type UserResponse struct {
